requirements.txt
flask
bson
google-generativeai
pymongo
authlib
